refactor(ui): extract message types out of model.go
model.go is the highest-churn file in the repo. Move the tea.Msg type
declarations and the small Cmd constructors that sit alongside them
(the action-menu view/copy cmds) into a new messages.go in the same
package. Mechanical cut — no renames, no signature changes, no
behaviour: the dispatcher and the modal-routing invariants stay in
model.go, deliberately out of scope. model.go drops ~160 lines.
Closes #60.
